Remove Reference from utilities which return types
