+++ src/transforms/v2-to-v3/ts-type/getV3ClientType.ts
@@ -6,9 +6,9 @@ import {
   V2_CLIENT_OUTPUT_SUFFIX_LIST,
 } from "../config";
 import { getDefaultLocalName } from "../utils";
-import { getTypeRefForString } from "./getTypeRefForString";
+import { getTypeForString } from "./getTypeForString";
 
-export interface GetV3ClientTypeReferenceOptions {
+export interface GetV3ClientTypeOptions {
   v2ClientLocalName: string;
   v2ClientName: string;
   v2ClientTypeName: string;
@@ -20,15 +20,15 @@ const getTypeRefWithV3ClientDefaultLocalName = (
   v3ClientTypeName: string
 ) => j.tsTypeReference(j.identifier([v3ClientDefaultLocalName, v3ClientTypeName].join(".")));
 
-export const getV3ClientTypeReference = (
+export const getV3ClientType = (
   j: JSCodeshift,
-  { v2ClientLocalName, v2ClientName, v2ClientTypeName }: GetV3ClientTypeReferenceOptions
+  { v2ClientLocalName, v2ClientName, v2ClientTypeName }: GetV3ClientTypeOptions
 ): TSType => {
   const clientTypesMap = CLIENT_TYPES_MAP[v2ClientName];
   const defaultLocalName = getDefaultLocalName(v2ClientLocalName);
 
   if (Object.keys(clientTypesMap).includes(v2ClientTypeName)) {
-    return getTypeRefForString(j, defaultLocalName, clientTypesMap[v2ClientTypeName]);
+    return getTypeForString(j, defaultLocalName, clientTypesMap[v2ClientTypeName]);
   }
 
   for (const inputSuffix of V2_CLIENT_INPUT_SUFFIX_LIST) {
