+++ src/transforms/v2-to-v3/ts-type/index.ts
@@ -1,4 +1,4 @@
 export * from "./getClientTypeNames";
-export * from "./getV3ClientTypeReference";
+export * from "./getV3ClientType";
 export * from "./getV3ClientTypesCount";
 export * from "./replaceTSQualifiedName";
